Move to tsc in compiling
