+++ package.json
@@ -9,22 +9,20 @@
     "node": ">= 16"
   },
   "scripts": {
-    "build": "parcel build --no-cache",
+    "test": "node scripts/npm-test.js",
     "type-check": "tsc --noEmit",
     "lint": "eslint .",
+    "test:unit": "jest",
     "test:app-src": "ts-node ./test/app/src/index.ts",
     "test:app-dist": "npm run build && node ./test/app/dist/index.js",
-    "test:app": "npm run test:app-src && npm run test:app-dist",
-    "test:unit": "jest",
-    "test": "node scripts/npm-test.js",
+    "build": "node scripts/npm-build.js",
+    "build:dist": "tsc -p tsconfig.dist.json",
     "prepare": "husky"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@parcel/packager-ts": "2.11.0",
-    "@parcel/transformer-typescript-types": "2.11.0",
     "@typescript-eslint/eslint-plugin": "7.0.2",
     "@typescript-eslint/parser": "7.0.2",
     "colors": "1.4.0",
@@ -32,7 +30,6 @@
     "eslint-config-google": "0.14.0",
     "husky": "9.0.11",
     "jest": "29.7.0",
-    "parcel": "2.11.0",
     "ts-jest": "29.1.2",
     "ts-node": "10.9.2",
     "typescript": "5.3.3",
