+++ scripts/npm-test.js
@@ -1,27 +1,12 @@
 const colors = require('colors/safe');
-const cp = require('node:child_process');
-const path = require('node:path');
+const {run, logTitle, cdToProj} = require('./common');
 
-process.chdir(path.join(__dirname, '..'));
-
+cdToProj();
 let wereErrors = false;
 
-function logTitle(title) {
-    console.log(colors.cyan(title));
-}
-
-function run(title, cmd) {
-    logTitle(title);
-    try {
-        cp.execSync(cmd, {stdio: 'inherit'});
-    } catch (e) {
-        wereErrors = true;
-    }
-}
-
-run('==> Running type check', 'npm run type-check');
-run('==> Running linters', 'npm run lint');
-run('==> Running unit tests', 'npm run test:unit');
+run('==> Running type check', 'npm run type-check') || (wereErrors = true);
+run('==> Running linters', 'npm run lint') || (wereErrors = true);
+run('==> Running unit tests', 'npm run test:unit') || (wereErrors = true);
 
 logTitle('==> Result');
 if (wereErrors) {
